+++ src/Exceptions/MimeTypesException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AwaisJameel\MimeTypes\Exceptions;
+
+use RuntimeException;
+
+abstract class MimeTypesException extends RuntimeException {}