+++ src/MimeTypesServiceProvider.php
@@ -9,13 +9,13 @@ class MimeTypesServiceProvider extends PackageServiceProvider
 {
     public function configurePackage(Package $package): void
     {
-        /*
-         * This class is a Package Service Provider
-         *
-         * More info: https://github.com/spatie/laravel-package-tools
-         */
         $package
             ->name('mimetypes')
             ->hasConfigFile();
+    }
+
+    public function packageRegistered(): void
+    {
+        $this->app->singleton(MimeTypes::class);
     }
 }