+++ tests/TestCase.php
@@ -3,20 +3,10 @@
 namespace AwaisJameel\MimeTypes\Tests;
 
 use AwaisJameel\MimeTypes\MimeTypesServiceProvider;
-use Illuminate\Database\Eloquent\Factories\Factory;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 class TestCase extends Orchestra
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'AwaisJameel\\MimeTypes\\Database\\Factories\\'.class_basename($modelName).'Factory'
-        );
-    }
-
     protected function getPackageProviders($app)
     {
         return [
@@ -26,11 +16,6 @@ class TestCase extends Orchestra
 
     public function getEnvironmentSetUp($app)
     {
-        config()->set('database.default', 'testing');
-
-        /*
-        $migration = include __DIR__.'/../database/migrations/create_mimetypes_table.php.stub';
-        $migration->up();
-        */
+        $app['config']->set('cache.default', 'array');
     }
 }